+++ src/extension/cell.ts
@@ -28,7 +28,7 @@ import {
   XTermState,
 } from './terminal/terminalState'
 import ContextState from './contextState'
-import { IRunnerEnvironment } from './runner'
+import { IRunnerEnvironment } from './runner/program'
 
 const NOTEBOOK_SELECTION_COMMAND = '_notebook.selectKernel'
