+++ src/extension/provider/runmeTask.ts
@@ -25,7 +25,7 @@ import getLogger from '../logger'
 import { asWorkspaceRelativePath, getAnnotations, getWorkspaceEnvs } from '../utils'
 import { Serializer, RunmeTaskDefinition } from '../../types'
 import { SerializerBase } from '../serializer'
-import type { IRunner, IRunnerEnvironment, RunProgramOptions } from '../runner'
+import type { IRunner, IRunnerEnvironment, RunProgramOptions } from '../runner/program'
 import { getCellCwd, getCellProgram, parseCommandSeq } from '../executors/utils'
 import { Kernel } from '../kernel'
 import RunmeServer from '../server/runmeServer'
